refactor: Move `Tk2Circuit` to its own file
